Added IN operator Separator attribute
I added a new attribute (#11) to enable the user to specify the separator character(s) between multiple values for the IN or Not IN conditional operators. This attribute is used by the JS manageFilter routine to convert that separator into the value that an Apex IG needs (01x).

I also changed the conditional operator labels and sort order to better match existing Apex operator lists.